Implement reformatting the scoresheet.
The input format is designed to be easy to check against the paper
forms we'll be filling in, and isn't very useful for processing.
Add a step which converts the input to a more useful layout.

This includes some placeholder functions which will contain the body
of the scoring logic.